+++ scripts/applications/craft/material-criteria-dialog.mjs
@@ -25,7 +25,8 @@ export default class MaterialCriteriaDialog extends Dialog5e {
     buttons: [
       { action: "add", label: "SIMPLE_SHOP_CRAFT_5E.RecipeEditor.AddMaterialCriteria", icon: "fas fa-plus", default: true }
     ],
-    form: { handler: MaterialCriteriaDialog.#onSubmit }
+    form: { handler: MaterialCriteriaDialog.#onSubmit },
+    onSubmit: null
   };
 
   /** @override */